+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+v3.1.0
+=======
+* Json body parameters no longer throws when:
+  * Content-type is omitted
+  * Content-type is any other than application/json
+
 v3.0.1
 =======
 * Fixed a bug related to specifications that always returned all available values. Fixed specs:
@@ -29,6 +35,12 @@ v3.0.0
   * `DateBetween` - use `Between`
   * `EqualEnum` - use `Equal`
   * `@Joins` container annotation - use repeated `@Join` and `@JoinFetch`
+
+v2.19.0
+=======
+* Json body parameters no longer throws when:
+  * Content-type is omitted
+  * Content-type is any other than application/json
 
 v2.18.1
 =======